Adds events.
