+++ financial-connections/src/main/java/com/stripe/android/financialconnections/features/institutionpicker/InstitutionPickerViewModel.kt
@@ -11,12 +11,12 @@ import com.airbnb.mvrx.ViewModelContext
 import com.stripe.android.core.Logger
 import com.stripe.android.financialconnections.FinancialConnectionsSheet
 import com.stripe.android.financialconnections.analytics.FinancialConnectionsAnalyticsTracker
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.FeaturedInstitutionsLoaded
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.InstitutionSelected
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.PaneLoaded
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.SearchScroll
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.SearchSucceeded
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.Error
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.FeaturedInstitutionsLoaded
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.InstitutionSelected
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.PaneLoaded
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.SearchScroll
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.SearchSucceeded
 import com.stripe.android.financialconnections.analytics.logError
 import com.stripe.android.financialconnections.domain.FeaturedInstitutions
 import com.stripe.android.financialconnections.domain.GetManifest
@@ -65,7 +65,7 @@ internal class InstitutionPickerViewModel @Inject constructor(
                     }
                 }.onFailure {
                     logger.error("Error fetching featured institutions", it)
-                    eventTracker.track(FinancialConnectionsEvent.Error(Pane.INSTITUTION_PICKER, it))
+                    eventTracker.track(Error(Pane.INSTITUTION_PICKER, it))
                 }
 
             val (institutions, duration) = result.getOrNull() ?: Pair(emptyList(), 0L)
