+++ financial-connections/src/main/java/com/stripe/android/financialconnections/features/linkstepupverification/LinkStepUpVerificationViewModel.kt
@@ -11,13 +11,13 @@ import com.airbnb.mvrx.Uninitialized
 import com.airbnb.mvrx.ViewModelContext
 import com.stripe.android.core.Logger
 import com.stripe.android.financialconnections.analytics.FinancialConnectionsAnalyticsTracker
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.Error
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.VerificationStepUpError
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.VerificationStepUpError.Error.ConsumerNotFoundError
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.VerificationStepUpError.Error.LookupConsumerSession
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.VerificationStepUpError.Error.MarkLinkVerifiedError
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.VerificationStepUpError.Error.StartVerificationError
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.VerificationStepUpSuccess
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.Error
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.VerificationStepUpError
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.VerificationStepUpError.Error.ConsumerNotFoundError
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.VerificationStepUpError.Error.LookupConsumerSession
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.VerificationStepUpError.Error.MarkLinkVerifiedError
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.VerificationStepUpError.Error.StartVerificationError
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.VerificationStepUpSuccess
 import com.stripe.android.financialconnections.domain.ConfirmVerification
 import com.stripe.android.financialconnections.domain.GetCachedAccounts
 import com.stripe.android.financialconnections.domain.GetManifest
