+++ financial-connections/src/main/java/com/stripe/android/financialconnections/features/networkinglinkverification/NetworkingLinkVerificationViewModel.kt
@@ -11,14 +11,14 @@ import com.airbnb.mvrx.Uninitialized
 import com.airbnb.mvrx.ViewModelContext
 import com.stripe.android.core.Logger
 import com.stripe.android.financialconnections.analytics.FinancialConnectionsAnalyticsTracker
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.Error
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.VerificationError
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.VerificationError.Error.ConsumerNotFoundError
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.VerificationError.Error.LookupConsumerSession
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.VerificationError.Error.NetworkedAccountsRetrieveMethodError
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.VerificationError.Error.StartVerificationSessionError
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.VerificationSuccess
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.VerificationSuccessNoAccounts
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.Error
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.VerificationError
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.VerificationError.Error.ConsumerNotFoundError
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.VerificationError.Error.LookupConsumerSession
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.VerificationError.Error.NetworkedAccountsRetrieveMethodError
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.VerificationError.Error.StartVerificationSessionError
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.VerificationSuccess
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.VerificationSuccessNoAccounts
 import com.stripe.android.financialconnections.domain.ConfirmVerification
 import com.stripe.android.financialconnections.domain.FetchNetworkedAccounts
 import com.stripe.android.financialconnections.domain.GetManifest
