+++ financial-connections/src/main/java/com/stripe/android/financialconnections/features/networkingsavetolinkverification/NetworkingSaveToLinkVerificationViewModel.kt
@@ -8,12 +8,12 @@ import com.airbnb.mvrx.Uninitialized
 import com.airbnb.mvrx.ViewModelContext
 import com.stripe.android.core.Logger
 import com.stripe.android.financialconnections.analytics.FinancialConnectionsAnalyticsTracker
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.Error
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.PaneLoaded
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.VerificationError
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.VerificationError.Error.ConfirmVerificationSessionError
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.VerificationError.Error.StartVerificationSessionError
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.VerificationSuccess
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.Error
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.PaneLoaded
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.VerificationError
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.VerificationError.Error.ConfirmVerificationSessionError
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.VerificationError.Error.StartVerificationSessionError
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.VerificationSuccess
 import com.stripe.android.financialconnections.domain.ConfirmVerification
 import com.stripe.android.financialconnections.domain.ConfirmVerification.OTPError
 import com.stripe.android.financialconnections.domain.GetCachedAccounts
