+++ financial-connections/src/main/java/com/stripe/android/financialconnections/presentation/FinancialConnectionsSheetNativeViewModel.kt
@@ -12,16 +12,14 @@ import com.airbnb.mvrx.PersistState
 import com.airbnb.mvrx.ViewModelContext
 import com.airbnb.mvrx.compose.mavericksActivityViewModel
 import com.stripe.android.core.Logger
-import com.stripe.android.financialconnections.FinancialConnections
-import com.stripe.android.financialconnections.FinancialConnectionsPublicEvent
 import com.stripe.android.financialconnections.FinancialConnectionsSheet
 import com.stripe.android.financialconnections.R
 import com.stripe.android.financialconnections.analytics.FinancialConnectionsAnalyticsTracker
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.AppBackgrounded
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.ClickNavBarBack
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.ClickNavBarClose
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.Complete
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.PaneLaunched
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.AppBackgrounded
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.ClickNavBarBack
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.ClickNavBarClose
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.Complete
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.PaneLaunched
 import com.stripe.android.financialconnections.di.APPLICATION_ID
 import com.stripe.android.financialconnections.di.DaggerFinancialConnectionsSheetNativeComponent
 import com.stripe.android.financialconnections.di.FinancialConnectionsSheetNativeComponent
@@ -307,9 +305,6 @@ internal class FinancialConnectionsSheetNativeViewModel @Inject constructor(
 
     fun onPaneLaunched(pane: Pane, referrer: Pane?) {
         viewModelScope.launch {
-            FinancialConnections.emitEvent(
-                FinancialConnectionsPublicEvent.Launched(pane.toString())
-            )
             eventTracker.track(
                 PaneLaunched(
                     referrer = referrer,
